core:  - renamed package resourceStatistics to statistics
